refactor: Inject dependencies into handler function
