+++ .babelrc
@@ -1,1 +1,5 @@
-{ "presets": ["es2015"] }
+{
+  "presets": [
+    "env",
+  ]
+}
