+++ package.json
@@ -9,19 +9,23 @@
   "author": "Patrick Simpson",
   "license": "ISC",
   "dependencies": {
-    "babel-cli": "^6.9.0",
-    "babel-preset-es2015": "^6.9.0",
+    "babel-env": "^2.4.1",
+    "lodash": "^4.13.1",
+    "node-env-file": "^0.1.8",
     "node-file-exists": "^1.1.0",
     "path": "^0.12.7",
     "slack": "^7.2.0",
-    "node-env-file": "^0.1.8",
-    "lodash": "^4.13.1",
     "slackbots": "^0.5.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.9.0"
+    "@babel/core": "^7.1.2",
+    "@babel/preset-env": "^7.1.0",
+    "babel-cli": "^6.26.0",
+    "eslint": "^5.7.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.14.0"
   },
   "engines": {
-    "node": "4.2.6"
+    "node": "10.5.0"
   }
 }
